Main menu and FXML removal
Good news! the emulator is now working. Bad news: The build won't even compile with the javafx code, as this is not a javaFX program. The main menu is designed, but I need someone to confirm it. Also removed the FXML files as they kept me from launching the app. I have a spare copy if they need to be put back in. I would've done more, but I have classes soon. I'll see you later today!